Decompose live runtime host service
