+++ desktop/runtime-host/live-runtime-service.cts
@@ -1,30 +1,14 @@
-import type { AgentMessage } from "@mariozechner/pi-agent-core";
 import type {
   ComposerAttachment,
-  ComposerSlashCommand,
   ComposerStateRequest,
   ComposerStreamingBehavior,
   ComposerThinkingLevel,
-  DesktopEvent,
 } from "../../shared/desktop-contracts.ts";
-import {
-  appSettingsSlashCommand,
-  compactSlashCommand,
-  parseCompactSlashCommand,
-} from "../../shared/composer-slash-commands.ts";
+import { parseCompactSlashCommand } from "../../shared/composer-slash-commands.ts";
 import { getDesktopWorkingDirectory } from "../../shared/desktop-working-directory.ts";
 import { createLocalThreadDraft, getPersistedSessionPath } from "../../shared/session-paths.ts";
-import { type SessionPathEntry, buildThreadHistorySlice } from "../../shared/thread-history.ts";
-import {
-  buildThreadData,
-  setThreadCompactingState,
-  setThreadStreamingState,
-} from "../../shared/thread-data.ts";
 import { getPiModule } from "../pi-module.cts";
-import {
-  bindHeadlessAgentSessionExtensions,
-  discoverHeadlessAgentSessionResources,
-} from "../runtime/agent-session-extensions.cts";
+import { discoverHeadlessAgentSessionResources } from "../runtime/agent-session-extensions.cts";
 import { buildComposerAttachmentPrompt } from "../runtime/attachments.cts";
 import {
   buildComposerQueueSnapshotKey,
@@ -39,478 +23,16 @@ import {
   clampThinkingLevel,
   getAvailableThinkingLevelsForModel,
 } from "../runtime/composer-state.cts";
-import type { PiRuntime, RuntimeThreadReason } from "../runtime/types.cts";
-
-type RuntimeRecord = {
-  runtimePromise: Promise<PiRuntime>;
-  disposeTimeout: ReturnType<typeof setTimeout> | null;
-};
-
-type RuntimeToolProgress = {
-  toolCallId: string;
-  toolName: string;
-  args?: unknown;
-  partialResult?: { content?: unknown };
-  isError?: boolean;
-  terminal?: boolean;
-};
-
-const RUNTIME_IDLE_TIMEOUT_MS = 15 * 60 * 1_000;
-const LIVE_THREAD_UPDATE_THROTTLE_MS = 50;
-
-const runtimeRecords = new Map<string, RuntimeRecord>();
-const runtimeMutationTails = new Map<string, Promise<void>>();
-const liveThreadUpdateTimers = new WeakMap<PiRuntime, ReturnType<typeof setTimeout>>();
-const liveToolProgressByRuntime = new WeakMap<PiRuntime, Map<string, RuntimeToolProgress>>();
-
-const builtinCommandNames = new Set([compactSlashCommand.name]);
-
-function mapSessionCommands(session: PiRuntime["session"]): ComposerSlashCommand[] {
-  const commands: ComposerSlashCommand[] = [appSettingsSlashCommand, compactSlashCommand];
-  const extensionCommandNames = new Set<string>();
-
-  for (const command of session.extensionRunner.getRegisteredCommands()) {
-    if (builtinCommandNames.has(command.invocationName)) {
-      continue;
-    }
-
-    extensionCommandNames.add(command.invocationName);
-    commands.push({
-      name: command.invocationName,
-      description: command.description,
-      source: "extension",
-      sourceInfo: command.sourceInfo,
-    });
-  }
-
-  for (const template of session.promptTemplates) {
-    if (builtinCommandNames.has(template.name) || extensionCommandNames.has(template.name)) {
-      continue;
-    }
-
-    commands.push({
-      name: template.name,
-      description: template.description,
-      source: "prompt",
-      sourceInfo: template.sourceInfo,
-    });
-  }
-
-  if (session.settingsManager.getEnableSkillCommands()) {
-    for (const skill of session.resourceLoader.getSkills().skills) {
-      const skillCommandName = `skill:${skill.name}`;
-      if (
-        builtinCommandNames.has(skillCommandName) ||
-        extensionCommandNames.has(skillCommandName)
-      ) {
-        continue;
-      }
-
-      commands.push({
-        name: skillCommandName,
-        description: skill.description,
-        source: "skill",
-        sourceInfo: skill.sourceInfo,
-      });
-    }
-  }
-
-  return commands;
-}
-
-let eventSink: ((event: DesktopEvent) => void) | null = null;
-
-export function setRuntimeHostEventSink(sink: (event: DesktopEvent) => void) {
-  eventSink = sink;
-}
-
-function emitDesktopEvent(event: DesktopEvent) {
-  eventSink?.(event);
-}
-
-function getLiveToolProgress(runtime: PiRuntime) {
-  let progress = liveToolProgressByRuntime.get(runtime);
-  if (!progress) {
-    progress = new Map();
-    liveToolProgressByRuntime.set(runtime, progress);
-  }
-  return progress;
-}
-
-function hasDisplayableToolContent(content: unknown): content is string | unknown[] {
-  if (typeof content === "string") return content.trim().length > 0;
-  if (!Array.isArray(content)) return false;
-  return content.some((part) => {
-    if (typeof part === "string") return part.trim().length > 0;
-    if (!part || typeof part !== "object") return false;
-    const record = part as Record<string, unknown>;
-    if (record.type === "image") return true;
-    return typeof record.text === "string" && record.text.trim().length > 0;
-  });
-}
-
-function getLiveToolProgressMessages(runtime: PiRuntime) {
-  const progress = liveToolProgressByRuntime.get(runtime);
-  if (!progress || progress.size === 0) return [] as AgentMessage[];
-  return [...progress.values()].map((entry) => {
-    const content = entry.partialResult?.content;
-    const displayContent = hasDisplayableToolContent(content)
-      ? content
-      : [
-          {
-            type: "text",
-            text: entry.terminal
-              ? entry.isError
-                ? `${entry.toolName} failed.`
-                : `${entry.toolName} finished.`
-              : `Running ${entry.toolName}...`,
-          },
-        ];
-    return {
-      role: "toolResult",
-      toolName: entry.toolName,
-      isError: Boolean(entry.isError),
-      content: displayContent,
-      timestamp: `tool-progress:${entry.toolCallId}`,
-    } as unknown as AgentMessage;
-  });
-}
-
-function rememberRuntimeToolProgress(runtime: PiRuntime, entry: RuntimeToolProgress) {
-  getLiveToolProgress(runtime).set(entry.toolCallId, entry);
-}
-
-function clearRuntimeToolProgress(
-  runtime: PiRuntime,
-  options: { toolCallId?: string; toolName?: string } = {},
-) {
-  const progress = liveToolProgressByRuntime.get(runtime);
-  if (!progress) return;
-  if (options.toolCallId) progress.delete(options.toolCallId);
-  else if (!options.toolName) progress.clear();
-  else {
-    for (const [toolCallId, entry] of progress) {
-      if (entry.toolName === options.toolName) {
-        progress.delete(toolCallId);
-        break;
-      }
-    }
-  }
-  if (progress.size === 0) liveToolProgressByRuntime.delete(runtime);
-}
-
-function normalizeThreadDataForReason(
-  thread: ReturnType<typeof buildThreadData>,
-  reason: RuntimeThreadReason,
-) {
-  if (reason === "compaction-start") return setThreadCompactingState(thread, true);
-  if (reason !== "end" && reason !== "compaction") return thread;
-  return setThreadCompactingState(setThreadStreamingState(thread, false), false);
-}
-
-function buildLiveThreadData(runtime: PiRuntime) {
-  const sessionPath = runtime.session.sessionFile;
-  if (!sessionPath) return null;
-  const streamingMessage = runtime.session.state.streamingMessage;
-  const historySlice = buildThreadHistorySlice(
-    [...(runtime.session.sessionManager.getBranch() as SessionPathEntry[])],
-    0,
-  );
-  const sourceMessages = [
-    ...historySlice.sourceMessages,
-    ...(streamingMessage ? [streamingMessage] : []),
-    ...getLiveToolProgressMessages(runtime),
-  ] as AgentMessage[];
-  return buildThreadData({
-    sessionPath,
-    sourceMessages,
-    previousMessageCount: historySlice.previousMessageCount,
-    isStreaming: runtime.session.isStreaming,
-    isCompacting: runtime.session.isCompacting,
-  });
-}
-
-async function publishThreadUpdate(runtime: PiRuntime, reason: RuntimeThreadReason) {
-  const sessionPath = runtime.session.sessionFile;
-  if (!sessionPath) return;
-  const liveThread = buildLiveThreadData(runtime);
-  if (!liveThread) return;
-  emitDesktopEvent({
-    type: "thread-update",
-    reason,
-    projectId: runtime.cwd,
-    threadId: runtime.session.sessionId,
-    sessionPath,
-    thread: normalizeThreadDataForReason(liveThread, reason),
-    composer: await buildComposerState(runtime, { includeContextUsage: reason !== "update" }),
-  });
-}
-
-function publishComposerUpdate(
-  composer: Awaited<ReturnType<typeof buildComposerState>>,
-  context: { projectId?: string | null; sessionPath?: string | null } = {},
-) {
-  emitDesktopEvent({
-    type: "composer-update",
-    composer,
-    projectId: context.projectId ?? null,
-    sessionPath: context.sessionPath ?? null,
-  });
-}
-
-function clearRuntimeDisposeTimeout(runtimeKey: string) {
-  const record = runtimeRecords.get(runtimeKey);
-  if (!record?.disposeTimeout) return;
-  clearTimeout(record.disposeTimeout);
-  record.disposeTimeout = null;
-}
-
-function suspendRuntimeDisposal(runtimeKey: string) {
-  clearRuntimeDisposeTimeout(runtimeKey);
-}
-
-function scheduleRuntimeDisposal(runtimeKey: string) {
-  const record = runtimeRecords.get(runtimeKey);
-  if (!record) return;
-  clearRuntimeDisposeTimeout(runtimeKey);
-  record.disposeTimeout = setTimeout(() => {
-    void (async () => {
-      const currentRecord = runtimeRecords.get(runtimeKey);
-      if (!currentRecord || currentRecord !== record) return;
-      try {
-        const runtime = await record.runtimePromise;
-        if (runtime.session.isStreaming || runtime.session.isCompacting) {
-          scheduleRuntimeDisposal(runtimeKey);
-          return;
-        }
-        runtime.session.dispose();
-      } catch {
-        // Ignore runtime disposal races.
-      } finally {
-        if (runtimeRecords.get(runtimeKey) === record) runtimeRecords.delete(runtimeKey);
-      }
-    })();
-  }, RUNTIME_IDLE_TIMEOUT_MS);
-}
-
-function cancelLiveThreadUpdate(runtime: PiRuntime) {
-  const timer = liveThreadUpdateTimers.get(runtime);
-  if (!timer) return;
-  clearTimeout(timer);
-  liveThreadUpdateTimers.delete(runtime);
-}
-
-function deferLiveThreadUpdate(runtime: PiRuntime, options: { requireStreaming?: boolean } = {}) {
-  cancelLiveThreadUpdate(runtime);
-  const timer = setTimeout(() => {
-    liveThreadUpdateTimers.delete(runtime);
-    if (options.requireStreaming !== false && !runtime.session.isStreaming) return;
-    void publishThreadUpdate(runtime, "update");
-  }, 0);
-  liveThreadUpdateTimers.set(runtime, timer);
-}
-
-function scheduleLiveThreadUpdate(runtime: PiRuntime) {
-  if (liveThreadUpdateTimers.has(runtime)) return;
-  const timer = setTimeout(() => {
-    liveThreadUpdateTimers.delete(runtime);
-    if (!runtime.session.isStreaming) return;
-    void publishThreadUpdate(runtime, "update");
-  }, LIVE_THREAD_UPDATE_THROTTLE_MS);
-  liveThreadUpdateTimers.set(runtime, timer);
-}
-
-async function createRuntime(options: {
-  cwd: string;
-  sessionManager?: PiRuntime["session"]["sessionManager"];
-}): Promise<PiRuntime> {
-  const {
-    AuthStorage,
-    ModelRegistry,
-    SessionManager,
-    SettingsManager,
-    createAgentSession,
-    getAgentDir,
-  } = await getPiModule();
-  const agentDir = getAgentDir();
-  const authStorage = AuthStorage.create();
-  const modelRegistry = ModelRegistry.create(authStorage, `${agentDir}/models.json`);
-  const settingsManager = SettingsManager.create(options.cwd, agentDir);
-  const sessionDir = settingsManager.getSessionDir() ?? undefined;
-  const { session } = await createAgentSession({
-    cwd: options.cwd,
-    agentDir,
-    authStorage,
-    modelRegistry,
-    settingsManager,
-    sessionManager: options.sessionManager ?? SessionManager.create(options.cwd, sessionDir),
-  });
-  const runtime = { cwd: options.cwd, session } satisfies PiRuntime;
-
-  session.subscribe((event) => {
-    const runtimeKey = getPersistedSessionPath(runtime.session.sessionFile);
-    if (runtimeKey) suspendRuntimeDisposal(runtimeKey);
-
-    if (event.type === "message_start") {
-      scheduleLiveThreadUpdate(runtime);
-      return;
-    }
-    if (event.type === "message_end") {
-      if (event.message.role === "user") {
-        cancelLiveThreadUpdate(runtime);
-        void publishThreadUpdate(runtime, "start");
-      } else {
-        if (event.message.role === "toolResult") {
-          const toolCallId = "toolCallId" in event.message ? event.message.toolCallId : undefined;
-          clearRuntimeToolProgress(runtime, {
-            toolCallId: typeof toolCallId === "string" ? toolCallId : undefined,
-            toolName: event.message.toolName,
-          });
-        }
-        deferLiveThreadUpdate(runtime, { requireStreaming: event.message.role === "toolResult" });
-      }
-      if (runtimeKey) scheduleRuntimeDisposal(runtimeKey);
-      return;
-    }
-    if (event.type === "agent_end") {
-      cancelLiveThreadUpdate(runtime);
-      void publishThreadUpdate(runtime, "end");
-      if (runtimeKey) scheduleRuntimeDisposal(runtimeKey);
-      return;
-    }
-    if (event.type === "compaction_start") {
-      cancelLiveThreadUpdate(runtime);
-      void publishThreadUpdate(runtime, "compaction-start");
-      void buildComposerState(runtime)
-        .then((composer) =>
-          publishComposerUpdate(composer, {
-            projectId: runtime.cwd,
-            sessionPath: runtime.session.sessionFile,
-          }),
-        )
-        .catch(() => {});
-      return;
-    }
-    if (event.type === "compaction_end") {
-      setTimeout(() => {
-        cancelLiveThreadUpdate(runtime);
-        void publishThreadUpdate(runtime, "compaction");
-        void buildComposerState(runtime)
-          .then((composer) =>
-            publishComposerUpdate(composer, {
-              projectId: runtime.cwd,
-              sessionPath: runtime.session.sessionFile,
-            }),
-          )
-          .catch(() => {});
-      }, 0);
-      return;
-    }
-    if (event.type === "message_update") {
-      scheduleLiveThreadUpdate(runtime);
-      return;
-    }
-    if (
-      event.type === "tool_execution_start" ||
-      event.type === "tool_execution_update" ||
-      event.type === "tool_execution_end"
-    ) {
-      rememberRuntimeToolProgress(runtime, {
-        toolCallId: event.toolCallId,
-        toolName: event.toolName,
-        args: "args" in event ? event.args : undefined,
-        partialResult:
-          event.type === "tool_execution_update"
-            ? event.partialResult
-            : event.type === "tool_execution_end"
-              ? event.result
-              : undefined,
-        isError: event.type === "tool_execution_end" ? event.isError : false,
-        terminal: event.type === "tool_execution_end",
-      });
-      scheduleLiveThreadUpdate(runtime);
-      return;
-    }
-    if (event.type === "queue_update") {
-      void buildComposerState(runtime)
-        .then((composer) =>
-          publishComposerUpdate(composer, {
-            projectId: runtime.cwd,
-            sessionPath: runtime.session.sessionFile,
-          }),
-        )
-        .finally(() => {
-          if (runtimeKey && !runtime.session.isStreaming) scheduleRuntimeDisposal(runtimeKey);
-        });
-    }
-  });
-
-  await bindHeadlessAgentSessionExtensions(session);
-  return runtime;
-}
-
-function registerRuntime(runtimeKey: string, runtimePromise: Promise<PiRuntime>) {
-  const record: RuntimeRecord = { runtimePromise, disposeTimeout: null };
-  runtimeRecords.set(runtimeKey, record);
-  return record;
-}
-
-function getCachedRuntimeForSessionPath(sessionPath: string) {
-  const persistedSessionPath = getPersistedSessionPath(sessionPath);
-  return persistedSessionPath
-    ? (runtimeRecords.get(persistedSessionPath)?.runtimePromise ?? null)
-    : null;
-}
-
-async function getOrCreateRuntimeForSessionPath(
-  sessionPath: string,
-  options: { suspendDisposal?: boolean } = {},
-) {
-  const persistedSessionPath = getPersistedSessionPath(sessionPath);
-  if (!persistedSessionPath)
-    throw new Error("A persisted session path is required to open a live runtime.");
-  const existingRuntime = runtimeRecords.get(persistedSessionPath);
-  if (existingRuntime) {
-    if (options.suspendDisposal) suspendRuntimeDisposal(persistedSessionPath);
-    return await existingRuntime.runtimePromise;
-  }
-  const { SessionManager } = await getPiModule();
-  const sessionManager = SessionManager.open(persistedSessionPath);
-  let record: RuntimeRecord | null = null;
-  const runtimePromise = createRuntime({ cwd: sessionManager.getCwd(), sessionManager }).catch(
-    (error) => {
-      if (record && runtimeRecords.get(persistedSessionPath) === record)
-        runtimeRecords.delete(persistedSessionPath);
-      throw error;
-    },
-  );
-  record = registerRuntime(persistedSessionPath, runtimePromise);
-  return runtimePromise;
-}
-
-async function createRuntimeForNewSession(cwd: string) {
-  const runtime = await createRuntime({ cwd });
-  const runtimeKey = getPersistedSessionPath(runtime.session.sessionFile);
-  if (runtimeKey) registerRuntime(runtimeKey, Promise.resolve(runtime));
-  return runtime;
-}
-
-async function withRuntimeMutationLock<T>(runtimeKey: string, task: () => Promise<T>) {
-  const previousTail = runtimeMutationTails.get(runtimeKey) ?? Promise.resolve();
-  let releaseCurrentTail: (() => void) | undefined;
-  const currentTail = new Promise<void>((resolve) => {
-    releaseCurrentTail = resolve;
-  });
-  const nextTail = previousTail.then(() => currentTail);
-  runtimeMutationTails.set(runtimeKey, nextTail);
-  await previousTail;
-  try {
-    return await task();
-  } finally {
-    releaseCurrentTail?.();
-    if (runtimeMutationTails.get(runtimeKey) === nextTail) runtimeMutationTails.delete(runtimeKey);
-  }
-}
+import type { PiRuntime } from "../runtime/types.cts";
+import { publishComposerUpdate, publishThreadUpdate } from "./live-thread-publisher.cts";
+import {
+  createRuntimeForNewSession,
+  getCachedRuntimeForSessionPath,
+  getOrCreateRuntimeForSessionPath,
+  scheduleRuntimeDisposal,
+  withRuntimeMutationLock,
+} from "./live-runtime-registry.cts";
+import { mapSessionCommands } from "./slash-command-service.cts";
 
 async function emitComposerUpdate(request: ComposerStateRequest = {}) {
   const persistedSessionPath = getPersistedSessionPath(request.sessionPath);
